bugs fixed

+++ src/main/java/it/tafaq/springboot/onlineshop/RestController/CategoryController.java
@@ -3,8 +3,14 @@ package it.tafaq.springboot.onlineshop.RestController;
 import it.tafaq.springboot.onlineshop.dto.BrandDto;
 import it.tafaq.springboot.onlineshop.dto.CategoryDto;
 import it.tafaq.springboot.onlineshop.entity.Category;
+import it.tafaq.springboot.onlineshop.entity.User;
 import it.tafaq.springboot.onlineshop.service.CategoryService;
+import it.tafaq.springboot.onlineshop.service.UserService;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,16 +20,28 @@ import java.util.Optional;
 @RequestMapping("/api/categories")
 public class CategoryController {
     private final CategoryService categoryService;
+    private final UserService userService;
 
-    public CategoryController(CategoryService categoryService) {
+    public CategoryController(CategoryService categoryService , UserService userService) {
         this.categoryService = categoryService;
+        this.userService = userService;
     }
 
     @PostMapping("/")
-    public void addNewCategory(@RequestBody CategoryDto categoryDto) {
+    public ResponseEntity<String> addNewCategory(@RequestBody CategoryDto categoryDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null ) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
+        String email = authentication.getName();
+        User currentUser = userService.findByEmail(email);
+        if (currentUser.getRole().equals("ROLE_USER") || currentUser.getRole().equals("ROLE_ADMIN")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Users cannot add Category");
+        }
         Category newCategory = new Category();
         newCategory.setName(categoryDto.getName());
         categoryService.save(newCategory);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Category created");
     }
 
     @Cacheable(value = "categories" , key = "'all_categories'")
@@ -32,16 +50,19 @@ public class CategoryController {
         return categoryService.findAll();
     }
 
-    @PostMapping("/{id}")
-    public void updateCategory(@PathVariable Long id, @RequestBody BrandDto brandDto) {
-        Optional<Category> category = Optional.ofNullable(categoryService.findById(id));
-        category.get().setName(brandDto.getName());
-        categoryService.save(category.orElse(null));
-    }
-
     @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable Long id) {
+    public ResponseEntity<String> deleteCategory(@PathVariable Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null ) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
+        String email = authentication.getName();
+        User currentUser = userService.findByEmail(email);
+        if (currentUser.getRole().equals("ROLE_USER") || currentUser.getRole().equals("ROLE_ADMIN")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Users cannot delete Category");
+        }
         categoryService.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body("Category deleted");
     }
 
     @GetMapping("/{id}")
